+++ components/engine_traits/src/engines.rs
@@ -1,8 +1,9 @@
 // Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
 
-use crate::engine::{KvEngine, RaftEngine};
+use crate::engine::KvEngine;
 use crate::errors::Result;
 use crate::options::WriteOptions;
+use crate::raft_engine::RaftEngine;
 
 #[derive(Clone, Debug)]
 pub struct Engines<K, R> {
