+++ components/engine_traits/src/lib.rs
@@ -323,6 +323,8 @@ mod options;
 pub use crate::options::*;
 pub mod range;
 pub use crate::range::*;
+mod raft_engine;
+pub use raft_engine::{CacheStats, RaftEngine, RaftLogBatch};
 
 // These modules need further scrutiny
 
@@ -330,7 +332,6 @@ pub mod metrics_flusher;
 pub use crate::metrics_flusher::*;
 pub mod compaction_job;
 pub mod util;
-pub use crate::engine::{RaftEngine, RaftLogBatch};
 pub use compaction_job::*;
 
 // FIXME: This should live somewhere else
