+++ components/engine_traits/src/metrics_flusher.rs
@@ -6,7 +6,7 @@ use std::sync::mpsc::{self, Sender};
 use std::thread::{Builder as ThreadBuilder, JoinHandle};
 use std::time::{Duration, Instant};
 
-use crate::engine::RaftEngine;
+use crate::raft_engine::RaftEngine;
 
 use crate::*;
