+++ components/raft_log_engine/Cargo.toml
@@ -7,33 +7,24 @@ edition = "2018"
 [features]
 protobuf-codec = [
   "keys/protobuf-codec",
-  "txn_types/protobuf-codec",
   "raft/protobuf-codec",
   "kvproto/protobuf-codec",
 ]
 
 prost-codec = [
   "keys/prost-codec",
-  "txn_types/prost-codec",
   "raft/prost-codec",
   "kvproto/prost-codec",
 ]
 
-jemalloc = ["rocksdb/jemalloc"]
-portable = ["rocksdb/portable"]
-sse = ["rocksdb/sse"]
-
 [dependencies]
 engine_traits = { path = "../engine_traits" }
+tikv_util = { path = "../tikv_util" }
 hex = "0.4"
 keys = { path = "../keys" }
 num_cpus = "1"
-prometheus = { version = "0.8", features = ["nightly", "push"] }
 slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debug"] }
 slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "0e23a5baff302a9d7bccd85f8f31e43339c2f2c1" }
-tikv_alloc = { path = "../tikv_alloc" }
-tikv_util = { path = "../tikv_util" }
-txn_types = { path = "../txn_types"}
 lazy_static = "1.4.0"
 time = "0.1"
 configuration = { path = "../configuration" }
@@ -45,20 +36,5 @@ kvproto = { git = "https://github.com/pingcap/kvproto.git", default-features = f
 raft = { version = "0.6.0-alpha", default-features = false }
 protobuf = "2"
 
-[dependencies.rocksdb]
-git = "https://github.com/tikv/rust-rocksdb.git"
-package = "rocksdb"
-features = ["encryption", "static_libcpp"]
-
-[dependencies.prometheus-static-metric]
-git = "https://github.com/tikv/rust-prometheus.git"
-rev = "fd122caa03de8e7e5e4fae9372583aebf19e39f6"
-
 [dependencies.raft-engine]
-git = "https://github.com/Little-Wallace/raft-engine.git"
-branch = "no-raft"
-
-[dev-dependencies]
-kvproto = { git = "https://github.com/pingcap/kvproto.git", default-features = false }
-rand = "0.7"
-toml = "0.4"
+git = "https://github.com/tikv/raft-engine.git"
